+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.4.2 ()
+* Added auto table creation to reduce installation issues.
+* Added the option to enable or disable the blips for police stations.
+* Allow two players forced into one vehicle
+
 ## 1.4.1 (12/06/2017)
 * Add CopAddAdmin, CopAdd, CopRem and CopRank rcon command
 * Warning message when performing copadd/coprem when whitelist is disabled